requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
numpy
duckdb
matplotlib
seaborn
langchain
langchain-core
langchain-community
langchain-groq
langgraph
langgraph-checkpoint-postgres
fastembed
redis
psycopg[binary]
psycopg_pool
langfuse
python-dotenv